// src/types/landing.ts
export interface LinkData {
    id: string
    title: string
    description: string
    ios_url: string
    android_url: string
    web_url: string
    logo_url: string
    show_ios: boolean
    show_android: boolean
    show_web: boolean
    screenshots: string[]
    rating: number
    review_count: number
}